+++ packages/astro/e2e/view-transitions.test.js
@@ -189,6 +189,23 @@ test.describe('View Transitions', () => {
 		).toEqual(3);
 	});
 
+	test('Declarative Shadow DOM elements are attached after transitions', async ({
+		page,
+		astro,
+	}) => {
+		const loads = collectLoads(page);
+
+		// Go to page 1
+		await page.goto(astro.resolveUrl('/one'));
+
+		// transition to DSD page
+		await page.click('#click-declarative-shadow-dom');
+		const host = page.locator('#dsd-host');
+		await expect(host, 'should have content').toHaveText('Welcome to the Shadow Realm!');
+
+		expect(loads.length, 'There should only be 1 page load').toEqual(1);
+	});
+
 	test('Stylesheets in the head are waited on', async ({ page, astro }) => {
 		// Go to page 1
 		await page.goto(astro.resolveUrl('/one'));
@@ -1263,6 +1280,18 @@ test.describe('View Transitions', () => {
 		});
 		// Submit the form
 		await page.click('#submit');
+	});
+
+	test('form with hash in url should still submit', async ({ page, astro }) => {
+		let navigated;
+		await page.goto(astro.resolveUrl('/form-with-hash#test'));
+		page.on('request', (request) => {
+			expect(request.method()).toBe('POST');
+			navigated = true;
+		});
+		// Submit the form
+		await page.click('#submit');
+		expect(navigated).toBe(true);
 	});
 
 	test('Route announcer is invisible on page transition', async ({ page, astro }) => {