+++ packages/astro/src/core/middleware/sequence.ts
@@ -1,6 +1,5 @@
 import type { MiddlewareHandler, RewritePayload } from '../../types/public/common.js';
 import type { APIContext } from '../../types/public/context.js';
-import { AstroCookies } from '../cookies/cookies.js';
 import { ForbiddenRewrite } from '../errors/errors-data.js';
 import { AstroError } from '../errors/index.js';
 import { getParams, type Pipeline } from '../render/index.js';
@@ -78,8 +77,8 @@ export function sequence(...handlers: MiddlewareHandler[]): MiddlewareHandler {
 						carriedPayload = payload;
 						handleContext.request = newRequest;
 						handleContext.url = new URL(newRequest.url);
-						handleContext.cookies = new AstroCookies(newRequest);
 						handleContext.params = getParams(routeData, pathname);
+						handleContext.routePattern = routeData.route;
 						setOriginPathname(
 							handleContext.request,
 							oldPathname,