+++ packages/astro/src/runtime/server/render/csp.ts
@@ -20,7 +20,7 @@ export function renderCspContent(result: SSRResult): string {
 		finalScriptHashes.add(`'${scriptHash}'`);
 	}
 
-	let directives = '';
+	let directives;
 	if (result.directives.length > 0) {
 		directives = result.directives.join(';') + ';';
 	}
@@ -38,5 +38,5 @@ export function renderCspContent(result: SSRResult): string {
 	const strictDynamic = result.isStrictDynamic ? ` 'strict-dynamic'` : '';
 	const scriptSrc = `script-src ${scriptResources} ${Array.from(finalScriptHashes).join(' ')}${strictDynamic};`;
 	const styleSrc = `style-src ${styleResources} ${Array.from(finalStyleHashes).join(' ')};`;
-	return `${directives} ${scriptSrc} ${styleSrc}`;
+	return [directives, scriptSrc, styleSrc].filter(Boolean).join(' ');
 }