+++ packages/astro/src/zod.ts
@@ -1,0 +1,5 @@
+import * as mod from 'zod';
+
+export * from 'zod';
+export { mod as z };
+export default mod;