+++ packages/astro/test/units/assets/fonts/implementations.test.js
@@ -14,10 +14,14 @@ import { createCachedFontFetcher } from '../../../../dist/assets/fonts/implement
 import { createCapsizeFontMetricsResolver } from '../../../../dist/assets/fonts/implementations/font-metrics-resolver.js';
 import { createFontTypeExtractor } from '../../../../dist/assets/fonts/implementations/font-type-extractor.js';
 import {
+	createBuildUrlProxyHashResolver,
+	createDevUrlProxyHashResolver,
+} from '../../../../dist/assets/fonts/implementations/url-proxy-hash-resolver.js';
+import {
 	createBuildUrlResolver,
 	createDevUrlResolver,
 } from '../../../../dist/assets/fonts/implementations/url-resolver.js';
-import { createSpyStorage, simpleErrorHandler } from './utils.js';
+import { createSpyStorage, fakeHasher, simpleErrorHandler } from './utils.js';
 
 describe('fonts implementations', () => {
 	describe('createMinifiableCssRenderer()', () => {
@@ -83,21 +87,33 @@ describe('fonts implementations', () => {
 			hash: 'xxx',
 			url: 'abc',
 			preload: null,
-			data: {},
+			data: {
+				weight: undefined,
+				style: undefined,
+				subset: undefined,
+			},
 			init: null,
 		});
 		dataCollector.collect({
 			hash: 'yyy',
 			url: 'def',
 			preload: { type: 'woff2', url: 'def' },
-			data: {},
+			data: {
+				weight: undefined,
+				style: undefined,
+				subset: undefined,
+			},
 			init: null,
 		});
 		dataCollector.collect({
 			hash: 'xxx',
 			url: 'abc',
 			preload: null,
-			data: {},
+			data: {
+				weight: undefined,
+				style: undefined,
+				subset: undefined,
+			},
 			init: null,
 		});
 
@@ -110,9 +126,36 @@ describe('fonts implementations', () => {
 		);
 		assert.deepStrictEqual(preloadData, [{ type: 'woff2', url: 'def' }]);
 		assert.deepStrictEqual(collectedFonts, [
-			{ hash: 'xxx', url: 'abc', data: {}, init: null },
-			{ hash: 'yyy', url: 'def', data: {}, init: null },
-			{ hash: 'xxx', url: 'abc', data: {}, init: null },
+			{
+				hash: 'xxx',
+				url: 'abc',
+				data: {
+					weight: undefined,
+					style: undefined,
+					subset: undefined,
+				},
+				init: null,
+			},
+			{
+				hash: 'yyy',
+				url: 'def',
+				data: {
+					weight: undefined,
+					style: undefined,
+					subset: undefined,
+				},
+				init: null,
+			},
+			{
+				hash: 'xxx',
+				url: 'abc',
+				data: {
+					weight: undefined,
+					style: undefined,
+					subset: undefined,
+				},
+				init: null,
+			},
 		]);
 	});
 
@@ -349,29 +392,30 @@ describe('fonts implementations', () => {
 	});
 
 	it('createDevUrlResolver()', () => {
-		assert.equal(
-			createDevUrlResolver({ base: 'base/_astro/fonts' }).resolve('xxx.woff2'),
-			'/base/_astro/fonts/xxx.woff2',
-		);
+		const resolver = createDevUrlResolver({ base: 'base/_astro/fonts' });
+		assert.deepStrictEqual(resolver.getCspResources(), []);
+		assert.equal(resolver.resolve('xxx.woff2'), '/base/_astro/fonts/xxx.woff2');
+		assert.deepStrictEqual(resolver.getCspResources(), ["'self'"]);
 	});
 
 	describe('createBuildUrlResolver()', () => {
 		const base = 'foo/_custom/fonts';
 
 		it('works with no assetsPrefix', () => {
-			assert.equal(
-				createBuildUrlResolver({ base, assetsPrefix: undefined }).resolve('abc.ttf'),
-				'/foo/_custom/fonts/abc.ttf',
-			);
+			const resolver = createBuildUrlResolver({ base, assetsPrefix: undefined });
+			assert.deepStrictEqual(resolver.getCspResources(), []);
+			assert.equal(resolver.resolve('abc.ttf'), '/foo/_custom/fonts/abc.ttf');
+			assert.deepStrictEqual(resolver.getCspResources(), ["'self'"]);
 		});
 
 		it('works with assetsPrefix as string', () => {
+			const resolver = createBuildUrlResolver({ base, assetsPrefix: 'https://cdn.example.com' });
+			assert.deepStrictEqual(resolver.getCspResources(), []);
 			assert.equal(
-				createBuildUrlResolver({ base, assetsPrefix: 'https://cdn.example.com' }).resolve(
-					'foo.woff',
-				),
+				resolver.resolve('foo.woff'),
 				'https://cdn.example.com/foo/_custom/fonts/foo.woff',
 			);
+			assert.deepStrictEqual(resolver.getCspResources(), ['https://cdn.example.com']);
 		});
 
 		it('works with assetsPrefix object', () => {
@@ -382,7 +426,7 @@ describe('fonts implementations', () => {
 					fallback: 'https://cdn.example.com',
 				},
 			});
-
+			assert.deepStrictEqual(resolver.getCspResources(), []);
 			assert.equal(
 				resolver.resolve('bar.woff2'),
 				'https://fonts.cdn.example.com/foo/_custom/fonts/bar.woff2',
@@ -391,6 +435,92 @@ describe('fonts implementations', () => {
 				resolver.resolve('xyz.ttf'),
 				'https://cdn.example.com/foo/_custom/fonts/xyz.ttf',
 			);
+			assert.deepStrictEqual(resolver.getCspResources(), [
+				'https://fonts.cdn.example.com',
+				'https://cdn.example.com',
+			]);
 		});
+	});
+
+	it('createBuildUrlProxyHashResolver()', () => {
+		const resolver = createBuildUrlProxyHashResolver({
+			hasher: fakeHasher,
+			contentResolver: {
+				resolve: (url) => url,
+			},
+		});
+		assert.equal(
+			resolver.resolve({
+				cssVariable: '--foo',
+				data: {
+					weight: undefined,
+					style: undefined,
+					subset: undefined,
+				},
+				originalUrl: 'whatever',
+				type: 'woff2',
+			}),
+			'whatever.woff2',
+		);
+		assert.equal(
+			resolver.resolve({
+				cssVariable: '--foo',
+				data: { weight: 400, style: 'italic', subset: 'latin' },
+				originalUrl: 'whatever',
+				type: 'woff2',
+			}),
+			'whatever.woff2',
+		);
+	});
+
+	it('createDevUrlProxyHashResolver()', () => {
+		const resolver = createDevUrlProxyHashResolver({
+			baseHashResolver: createBuildUrlProxyHashResolver({
+				hasher: fakeHasher,
+				contentResolver: {
+					resolve: (url) => url,
+				},
+			}),
+		});
+		assert.equal(
+			resolver.resolve({
+				cssVariable: '--foo',
+				data: {
+					weight: undefined,
+					style: undefined,
+					subset: undefined,
+				},
+				originalUrl: 'whatever',
+				type: 'woff2',
+			}),
+			'foo-whatever.woff2',
+		);
+		assert.equal(
+			resolver.resolve({
+				cssVariable: '--foo',
+				data: { weight: 400, style: 'italic', subset: 'latin' },
+				originalUrl: 'whatever',
+				type: 'woff2',
+			}),
+			'foo-400-italic-latin-whatever.woff2',
+		);
+		assert.equal(
+			resolver.resolve({
+				cssVariable: '--foo',
+				data: { weight: '500', style: 'italic', subset: 'latin-ext' },
+				originalUrl: 'whatever',
+				type: 'woff2',
+			}),
+			'foo-500-italic-latin-ext-whatever.woff2',
+		);
+		assert.equal(
+			resolver.resolve({
+				cssVariable: '--foo',
+				data: { weight: [100, 900], style: 'italic', subset: 'cyrillic' },
+				originalUrl: 'whatever',
+				type: 'woff2',
+			}),
+			'foo-100-900-italic-cyrillic-whatever.woff2',
+		);
 	});
 });