+++ packages/db/src/core/schemas.ts
@@ -84,6 +84,7 @@ const textColumnBaseSchema = baseColumnSchema
 	.extend({
 		default: z.union([z.string(), sqlSchema]).optional(),
 		multiline: z.boolean().optional(),
+		enum: z.tuple([z.string()]).rest(z.string()).optional(), // At least one value required,
 	})
 	.and(
 		z.union([
@@ -94,7 +95,7 @@ const textColumnBaseSchema = baseColumnSchema
 			z.object({
 				// text primary key allows NULL values.
 				// NULL values bypass unique checks, which could
-				// lead to duplicate URLs per record in Astro Studio.
+				// lead to duplicate URLs per record.
 				// disable `optional` for primary keys.
 				primaryKey: z.literal(true),
 				optional: z.literal(false).optional(),