+++ packages/integrations/sitemap/src/index.ts
@@ -20,6 +20,7 @@ export type SitemapOptions =
 	| {
 			filenameBase?: string;
 			filter?(page: string): boolean;
+			customSitemaps?: string[];
 			customPages?: string[];
 
 			i18n?: {
@@ -38,6 +39,14 @@ export type SitemapOptions =
 			serialize?(item: SitemapItem): SitemapItem | Promise<SitemapItem | undefined> | undefined;
 
 			xslURL?: string;
+
+			// namespace configuration
+			namespaces?: {
+				news?: boolean;
+				xhtml?: boolean;
+				image?: boolean;
+				video?: boolean;
+			};
 	  }
 	| undefined;
 
@@ -88,7 +97,7 @@ const createPlugin = (options?: SitemapOptions): AstroIntegration => {
 
 					const opts = validateOptions(config.site, options);
 
-					const { filenameBase, filter, customPages, serialize, entryLimit } = opts;
+					const { filenameBase, filter, customPages, customSitemaps, serialize, entryLimit } = opts;
 
 					const outFile = `${filenameBase}-index.xml`;
 					const finalSiteUrl = new URL(config.base, config.site);
@@ -166,6 +175,7 @@ const createPlugin = (options?: SitemapOptions): AstroIntegration => {
 						}
 					}
 					const destDir = fileURLToPath(dir);
+					const lastmod = opts.lastmod?.toISOString();
 					const xslURL = opts.xslURL ? new URL(opts.xslURL, finalSiteUrl).href : undefined;
 					await writeSitemap(
 						{
@@ -175,7 +185,10 @@ const createPlugin = (options?: SitemapOptions): AstroIntegration => {
 							publicBasePath: config.base,
 							sourceData: urlData,
 							limit: entryLimit,
+							customSitemaps,
 							xslURL: xslURL,
+							lastmod,
+							namespaces: opts.namespaces,
 						},
 						config,
 					);