+++ packages/integrations/sitemap/src/write-sitemap.ts
@@ -4,8 +4,7 @@ import { normalize, resolve } from 'node:path';
 import { pipeline, Readable } from 'node:stream';
 import { promisify } from 'node:util';
 import type { AstroConfig } from 'astro';
-
-import { SitemapAndIndexStream, SitemapStream } from 'sitemap';
+import { SitemapAndIndexStream, SitemapIndexStream, SitemapStream } from 'sitemap';
 import replace from 'stream-replace-string';
 import type { SitemapItem } from './index.js';
 
@@ -13,11 +12,19 @@ type WriteSitemapConfig = {
 	filenameBase: string;
 	hostname: string;
 	sitemapHostname?: string;
+	customSitemaps?: string[];
 	sourceData: SitemapItem[];
 	destinationDir: string;
 	publicBasePath?: string;
 	limit?: number;
 	xslURL?: string;
+	lastmod?: string;
+	namespaces?: {
+		news?: boolean;
+		xhtml?: boolean;
+		image?: boolean;
+		video?: boolean;
+	};
 };
 
 // adapted from sitemap.js/sitemap-simple
@@ -29,8 +36,11 @@ export async function writeSitemap(
 		sourceData,
 		destinationDir,
 		limit = 50000,
+		customSitemaps = [],
 		publicBasePath = './',
 		xslURL: xslUrl,
+		lastmod,
+		namespaces = { news: true, xhtml: true, image: true, video: true },
 	}: WriteSitemapConfig,
 	astroConfig: AstroConfig,
 ) {
@@ -43,6 +53,13 @@ export async function writeSitemap(
 			const sitemapStream = new SitemapStream({
 				hostname,
 				xslUrl,
+				// Custom namespace handling
+				xmlns: {
+					news: namespaces?.news !== false,
+					xhtml: namespaces?.xhtml !== false,
+					image: namespaces?.image !== false,
+					video: namespaces?.video !== false,
+				},
 			});
 			const path = `./${filenameBase}-${i}.xml`;
 			const writePath = resolve(destinationDir, path);
@@ -64,11 +81,20 @@ export async function writeSitemap(
 				stream = sitemapStream.pipe(createWriteStream(writePath));
 			}
 
-			return [new URL(publicPath, sitemapHostname).toString(), sitemapStream, stream];
+			const url = new URL(publicPath, sitemapHostname).toString();
+			return [{ url, lastmod }, sitemapStream, stream];
 		},
 	});
 
 	const src = Readable.from(sourceData);
 	const indexPath = resolve(destinationDir, `./${filenameBase}-index.xml`);
+	for (const url of customSitemaps) {
+		SitemapIndexStream.prototype._transform.call(
+			sitemapAndIndexStream,
+			{ url, lastmod },
+			'utf8',
+			() => {},
+		);
+	}
 	return promisify(pipeline)(src, sitemapAndIndexStream, createWriteStream(indexPath));
 }